{{cookiecutter.project_dir}}/requirements/_base.txt
django==1.9.7

requests==2.10.0

# Config
django-environ==0.4.0
django-secure==1.0.1


# Database
psycopg2==2.6.1
django-model-utils==2.4


# Model
django-autoslug==1.9.3
unicode-slugify==0.1.3


# Form
django-bootstrap3==7.0.1
django-crispy-forms==1.6.0


# Images
easy-thumbnails==2.3
Pillow==3.2.0


# Redis
django-redis==4.4.3
redis>=2.10.0


# Static
django_compressor==2.0


# Time
pytz==2016.4
